# src/domain/entities/location.ts
export interface Location {
  location_id?: string;
  lat: number;
  lng: number;
  lat_dms: string;
  lng_dms: string;
  address: string;
  city: string;
  state: string;
  country: string;
  zip: string;
}